+++ source/server/http/admin.cc
@@ -176,10 +176,26 @@ bool filterParam(Http::Utility::QueryParams params, Buffer::Instance& response,
   return true;
 }
 
+// Helper method that returns true if the "list_keys" parameter is present.
+// The value is not kept and is assumed to be equivalent to true if the key exists.
+bool hasListKeysParam(Http::Utility::QueryParams params) {
+  return (params.find("list_keys") != params.end());
+}
+
+// Helper method to get a query parameter
+absl::optional<std::string> queryParam(Http::Utility::QueryParams params, std::string key) {
+  return (params.find(key) != params.end()) ? absl::optional<std::string>{params.at(key)}
+                                            : absl::nullopt;
+}
+
 // Helper method to get the format parameter
 absl::optional<std::string> formatParam(Http::Utility::QueryParams params) {
-  return (params.find("format") != params.end()) ? absl::optional<std::string>{params.at("format")}
-                                                 : absl::nullopt;
+  return queryParam(params, "format");
+}
+
+// Helper method to get the key parameter
+absl::optional<std::string> keyParam(Http::Utility::QueryParams params) {
+  return queryParam(params, "key");
 }
 
 // Helper method that ensures that we've setting flags based on all the health flag values on the
@@ -529,11 +545,37 @@ Http::Code AdminImpl::handlerClusters(absl::string_view url, Http::HeaderMap& re
   return Http::Code::OK;
 }
 
-// TODO(jsedgwick) Use query params to list available dumps, selectively dump, etc
-Http::Code AdminImpl::handlerConfigDump(absl::string_view, Http::HeaderMap& response_headers,
-                                        Buffer::Instance& response, AdminStream&) const {
-  envoy::admin::v2alpha::ConfigDump dump;
+void AdminImpl::writeDumpKeysAsJson(Envoy::Buffer::Instance& response) const {
+  envoy::admin::v2alpha::ConfigDumpKeys keys;
   for (const auto& key_callback_pair : config_tracker_.getCallbacksMap()) {
+    keys.add_keys(key_callback_pair.first);
+  }
+  response.add(MessageUtil::getJsonStringFromMessage(keys, true)); // pretty-print
+}
+
+Http::Code AdminImpl::handlerConfigDump(absl::string_view url, Http::HeaderMap& response_headers,
+                                        Buffer::Instance& response, AdminStream&) const {
+  Http::Utility::QueryParams query_params = Http::Utility::parseQueryString(url);
+  if (hasListKeysParam(query_params)) {
+    writeDumpKeysAsJson(response);
+    return Http::Code::OK;
+  }
+
+  auto callbacks_map = config_tracker_.getCallbacksMap();
+  const auto desired_key = keyParam(query_params);
+  if (desired_key.has_value()) {
+    const auto iter = callbacks_map.find(desired_key.value());
+    if (iter == callbacks_map.end()) {
+      response.add(
+          fmt::format("Key \"{}\" not found. Try /config_dump?list_keys to view available keys",
+                      desired_key.value()));
+      return Http::Code::NotFound;
+    }
+    callbacks_map = {{iter->first, iter->second}};
+  }
+
+  envoy::admin::v2alpha::ConfigDump dump;
+  for (const auto& key_callback_pair : callbacks_map) {
     ProtobufTypes::MessagePtr message = key_callback_pair.second();
     RELEASE_ASSERT(message, "");
     auto& any_message = *(dump.add_configs());