+++ docs/root/intro/version_history.rst
@@ -4,6 +4,7 @@ Version history
 1.13.0 (pending)
 ================
 * access log: added FILTER_STATE :ref:`access log formatters <config_access_log_format>` and gRPC access logger.
+* admin: added the ability to filter :ref:`/config_dump <operations_admin_interface_config_dump>` by key.
 * api: remove all support for v1
 * buffer: remove old implementation
 * build: official released binary is now built against libc++.